requirements.txt
apache-airflow==2.7.3
apache-airflow-providers-postgres==5.10.1
apache-airflow-providers-http==4.7.1
pandas==2.1.3
sqlalchemy==2.0.23
